fix darwin builds
This stubs out or tags things that weren't building on Darwin.
It may be a bit too heavy handed, although the only other platform
likely to be impacted would be Plan 9, and I don't see how these values
would have worked there either.
