+++ cmds/cpud/main_darwin.go
@@ -1,0 +1,16 @@
+// Copyright 2018-2019 the u-root Authors. All rights reserved
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+//go:build darwin
+// +build darwin
+
+package main
+
+import (
+	"fmt"
+)
+
+func main() {
+	fmt.Println("No support on Darwin right now")
+}
